Solution to Input and Print a Float

+++ Input and Print a Float@@6788d5012db25efec946f93a/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 
 int main() {
-    int a;
-    scanf("%d",&a)
-    printf("%d",a)
+    float a;
+    scanf("%f",&a)
+    printf("%f",a)
 
     return 0;
 }
